WIP: Static templates
Signed-off-by: Ritam Dey <esyray15@gmail.com>

Working on splitting the template into static PHP templates that hold
some boilerplate code. Completed the work on index.html.php.

The templates/header.html.php contains all the boilerplate code to start
a session, a database connection and the HTML navbar. It needs to be
included before any PHP

Te templates/footer.html.php contains all the boilerplate code to close
a session, a database session and HTML footer. It needs to included at
the end of all PHP