FIX BUGS: On a lot of places
Signed-off-by: Ritam <esyray15@gmail.com>

Fix undefined PHPSESSID index error on authentication/login.php, templates/header.html.php.
Fix argument error on checkActiveSession() in lib/sanitizers_validators.php
Deleted the unused <style> and added the <a> tag on the add to cart button.
Fix error of wrong function called as validateParameter() instead of validateISBN()
Moved all the script and CSS loading to the button of the HTML page to avoid slowing the loading of website