FIX BUG: Session not correctly resumed due to error in /lib/sanitizers_validators.php
Signed-off-by: Ritam <esyray15@gmail.com>

Author:    Ritam Dey <deyritam27031999@gmail.com>
Changes to be committed:
	modified:   lib/sanitizers_validators.php
A error in lib/sanitizers_validators.php occured due to calculation of timediff in SQL. This caused problems in PHP as DateTime can't create object when hh > 24. To workaround, now the started time is fetched as-is from database and the calculation of timediff is now handled in  PHP code. Also the session lifetime is now 5 days, instead of 15 days